requirements.txt
aiogram
python-dotenv
opencv-python
pillow
pytesseract
natasha
regex
